From Dave
Changes to module_io_domain.F in open_w_dataset.  The actual argument for the input dummy variable sysdepinfo is usually a character constant.  When we modify it with an internal write, we are clobbering memory.  The real program does not run on the IBM for some moving nested cases with certain physics options (PX surface for example, but I think that is just happenstance and not intrinsically related to this physics option in particular). 

John

1. Workaround for PGI compiler bug

There arose an internal namespace collision from the name of a new module added last week.  Portland Group promises to fix. Meanwhile, a workaround: just change the name slightly so it doesn't collide with the internal symbol that was generated for PRIVATE storage of some module variables.

2.  Fix for JM fat-finger bug that caused an error under g95.

Added a new routine last week that called an instance of a module procedure when it should be calling the module procedure itself.  Changed that and then corrected the ONLY clause to include the module procedure (also was missing in original commit).

3.  Workaround to get rid of anticipated "too complicated" msgs.

This is both workaround and just general cleanup too.  Those new linked-lists of all the state variables that were added to allow run time configuration of I/O can also be used to replace the registry-generated shifts in X and Y for moving nests.  The registry-generated shifts were verging on too complicated.  I had intended to implement the shifts this new way anyway. NMM developers, please note: this is ARW-only. I haven't touched the NMM shifts, they're still being done the old way.


4.  Another workaround to get rid of anticipated "too complicated" msgs.

In the same spirit as #3, above.  It turns out that the registry-generated halo routine used for moving-nest shifts is exactly the same whether for X or Y shifting.  So why compile both? Especially since these are among the worst compiler chokers in the code right now.  So, have modified code to use same halo exchanges for both X and Y directions and have modified registry program to only generate the one set.  As above, I only did this for the ARW moving nest code; NMM moving nest code is untouched.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@4166 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d