1) Add ROS subgrid variable to the "state:..." clause in package declaration.
This fixes a situation where a 2D field with zero length dimensions was being written to WRF I/O API.  The dimensions were zero length because ROS is a subgrid-dimensioned variable but the subgrid dimensions are zero unless fire is turned on.  This caused problems for parallel NetCDF.  (Jan and Jonathan, can you check and make sure this change is okay for you, please?)  Also including a fix for netcdf and parallel NetCDF implementations of I/O API so that they handle this situation correctly (see below). 

2) Fix for CAM threading bug

A few variables defined at the top of module_ra_cam need to be declared as automatic variables inside the radcswmx routine instead for thread-safety.  This was causing very ugly failures for the NRCM runs on the Argonne BG/P, and would also cause problems for other OpenMP runs with CAM.  (Hoping we notice the set of CAM OpenMP failures in the regression tests go away too.)  Also, there is some non-thread-safe code in module_radiation_driver that I have moved up and out of the OMP loop.

3) Fix for threading error in sst_skin updating code.

This fell out in testing of NRCM on the ANL BG/P machine.  The new call to sst_skin_update needed to be inside the loop over num_tiles.

4) Fix for some missing SELECT cases for reading through pNetCDF.

Was not reading properly when the input stream was an auxinput.  Only parallel-NetCDF affected.

5) Performance mods to parallel NetCDF.

Do a parallel transpose before writing the data to the parallel file system.  Should provide better I/O performance by blocking the data.  Basically does an x-transpose (all points in x in processor) then writes to parallel-NetCDF.

6) fix problem with zero-length write requests in netcdf and pnetcdf

Added a function that checks and just returns, ignoring requests to write fields with zero length dimensions in calls to *_write_field. This can happen with subgrid arrays (like ROS, above) if the subgrid dimensions are not specfied (ie. if fire is not turned on in this case, though previous proposal will avoid that situation).

7) Change to clean command.

Added an -aa option that does what clean does but without deleting my stuff in the test/em_directory.

8) Restore regtest.csh. 

I accidentally changed this a couple commits ago. Steve noticed and gave me hell :-)
Just restoring to original state.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@4171 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d