1) Suppress compiler warnings in C code for registry (Bill. Gustafson at PNNL via Steve).
2) Change pointers to allocatable in domain data structure Roman Dubtsov
at Intel suggested this change.  It can be turned on and off using
-DUSE_ALLOCATABLES. Not sure yet what the default should be.  He claims
a 4% performance improvement. Here's his explanation of what it does:
"The difference is not really in the performance of ALLOCATE itself, it is
in the optimization of loops containing these variables. The problem with
pointers is that they can be modified to point to something else, so it's
much harder for the compiler to disambiguate (i.e., to be sure that there
are no aliases or dependencies due to a memory location being accessed in
more than one way). In contrast, an allocatable array always references
the memory that was allocated to it; consequently, it's much easier for
the compiler to prove that vectorization and other loop optimizations
are safe when allocatable arrays are used instead of pointers."

3) Put a wrapper around malloc to align data accesses Another one
from Roman.  This one is enabled by -DWRAP_MALLOC and the idea is
to have malloc return storage in aligned chunks.  The rationale from
Roman is as follows: 

"This patch makes results from identically configured runs bit-to-bit
identical with default or even relaxed precision settings. The patch
was verified w/ WRF3.0.1.1 and CONUS2.5km workload. Rationale behind
the patch is that default malloc() is guaranteed to return results that
are 8/16-byte aligned on IA for x86/x86_64, and results of certain
optimized loops depend on input data alignment (prologues/epilogues
after/before vectorized body have different trip counts). This would
not have mattered if malloc() that is called from ALLOCATE statements
was deterministic. However, many modern systems have ASLR (address space
layout randomization) enabled, so malloc() is not deterministic. Even w/o
ASLR, some MPIs (Intel MPI is among then) and Intel OpenMP library use
environment variables to communicate some data to profiling tools/etc,
which causes additional calls to realloc() and malloc() to be made. Since
values of environment variables vary from run to run, this makes malloc()
behavior non-deterministic from program's point of view. So, the patch
works around malloc() non-determinism by providing aligned addresses
that make prologue/epilogue trip counts constant."

What I think he's saying is that Intel's implementation of malloc is
introducing some otherwise uncontrollable perturbations with optimization
that causes the code to not give the same answers twice. Propose including 
this as an option, the way Roman has it, but default turned off.

5.  Wire Windows port for correct setting without user having to
set Windows needs nocolons = .true. always.  Also, it doesn't handle
redirection of unbuffered output to stderr properly. It buffers the
output, so that when an error occurs near beginning of run, the user
doesn't get any output at all.  This Windows-only change will have the
stderr from task 0 go to screen but all other tasks will still redirect
to files.

6.  Remove a comment about over-decomposition from the configure.wrf
files when they're generated.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@4193 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d