1) Add -j argument to the compile script
At Pete Johnsen's (Cray) request, adding a -j option to the compile script to control how many tasks are used for parallel make.  This was already controllable by setting the shell environment variable "J", but -j is simpler to use and document.  Added several lines of usage information to the output of "compile -h", including a line for the -j option.  Affected file is the compile script. 

2) PGI needs all files compiled -mp for smpar

After working with PGI, it appears the problem we've been seeing with bit-reproducibility with OpenMP has been because we don't compile all files with OpenMP.  The proposed commit modifies the options for PGI compiles to use both "-mp" and "-Mrecursive" for every Fortran source file, regardless of whether it contains OpenMP directives or not.  This eliminats 1-vs-4 thread bit-for-bit differences on Linux.  The MacOS systems still have differences -- continuing to explore that with PGI, they think it might be compiler bug on MacOS but still investigating.  Affected file:

    arch/configure_new.defaults

3) Fix a format statement in the parallel NetCDF I/O API

Changed 'i3' to '(i3)' in the format part of an internal write statement.  One of the compilers somewhere (ANL, I think) caught this.

    external/io_pnetcdf/wrf_io.F90

4) Reduce memory requirements of PBL driver

The PBL driver module allocates 10 local 3D arrays that are associated with the diff3d routine (idiff = 1), which is never turned on.  Change these to ALLOCATABLE and then allocate them only if idiff has been set to 1.  Also declared intent of idiff arguments in Boulac and MYJURB.

   phys/module_pbl_driver.F
   phys/module_bl_boulac.F
   phys/module_bl_myjurb.F

5) Disable transposes in parallel NetCDF I/O API

This disables a feature that was added to the parallel NetCDF I/O API a week or so ago (2010-02-19).  The parallel transposes that were added don't improve the I/O performance that much; plus, they cause problems on large numbers of processors when some processes end up having now data to write in the transposed state.  Needs a little more debugging; meantime, since we're up against a release, just disable the feature and go back to the old way of doing it (write to pNetCDF directly without trying t reorder the data for better blocking).

  external/io_pnetcdf/wrf_io.F90

6) Fix instance of hard-coded number of streams left over from earlier days

Code in output_wrf to deal with a topo flag checked only the first 11 history output streams.  Modify so that it checks all the streams now that number has been expanded.

  share/output_wrf.F

7) Added time manager Alarm information to restart files so these get set properly on restart

Alarms for input and output streams that are set on intervals other than the restart interval had been causing problems on restart.  The solution is to put the alarm information into the restart file too.  One caveat is that the number of I/O streams in the code that is restarting must equal the number of I/O streams in the code that wrote the restart.  That is, both runs must have had MAX_HISTORY the same in the configure.wrf file when the were compiled.  If they don't match, or if the restart dataset does not contain the alarm information, the code will still restart, but this new feature won't work. The restarting code will just ignore the Alarm information in the restart file.  This involved a couple changes in the stand-alone time manager code (external/esmf_time_f90) too, to include the correct logic in the ClockAdvance code for when an Alarm has both RingInterval and RingTime set, as would be the case for an alarm that starts with a RingTime that appears to be off-interval relative to restart time (but correct relative to start of simulation).

  share/input_wrf.F
  share/output_wrf.F
  external/esmf_time_f90/ESMF_Clock.F90

8) Added GPU-accelerator directive versions of WSM3 and WSM5 (from Craig Toepfer at PGI)

PGI version 10 compilers include accelerator directives for GPU.  Craig Toepfer at PGI has developed GPU versions of the 3.1.1 WSM5 and WSM6 microphysics; including those here as separate files added to repository.  They will be selected in our out at compile time.  These are 3.1.1 versions.  There hasn't been a chance to update these to the v3.2 versions of the WSM3 and WSM5 microphysics (nor the hand-coded GPU version of WSM5 that's on the WRF GPU web page).  Everything is #ifdef _ACCEL and there is no effect on other builds of code.  New diffs file with these changes included.

  arch/configure_new.defaults  (add a build option to use the accel directives)
  main/module_wrf_top.F        (code to use multiple GPUs with OpenMP)
  tools/standard.c             (to not remove as comments the accel directives)
  phys/module_mp_wsm3_accel.F  (new file)
  phys/module_mp_wsm5_accel.F  (new file)




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@4241 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d