Some small commits prior to the first friendly release.  These are semi-emergency mods from Dave for the pre-release this week. Only one of these mods (the first below) should be reviewed by the physics folks at a later date to get a better solution.
1) The radiation driver fails with 8-byte reals.  There are a 
few wrf_debug statements that fix that.  This should be addressed
permanently in another fashion.
M    phys/module_radiation_driver.F

2) A new regtest.csh is available in the tools dir.  This has
the new physics tested, as well as dropping out the tests for
failed #9 (Milbrandt) and #12 (urban_opt =3).  For the large 
weekly regtests, these tests are still conducted.
M    tools/regtest.csh

3) The pre-release name is now "V3.3pre#1 " 
M    inc/version_decl

4) The gfortran code requires a larger space for the random
number seed (it was two).  The space was originally allocated
for 18 integers (which has been retained), but now all 18 
spaces are available for the seed computation.  This can be a 
permanent fix as is.  However, there are a number of physics
routines where we should consistently:
a) make the seed an allocatable array
b) find out how big to make it on this compiler
c) allocate the space for the seed
d) use the seed
e) deallocate the space at the end of the routine
M    dyn_em/module_stoch.F



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@4671 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d