Two changes that impact only ARW chemistry runs.
1)  Make the time averaging variables used in the online/offline chemistry runs 
optional in the chemics_init.F routine. Also added a message to the chemistry 
initialization routine to tell the users when these chemistry time averaged 
variables are not included.

2) Fix an issue with the chem KPP compile.  The BUILD_CHEM=1 option was left off 
of a line in the chem KPP compile script (compile_wkc).  This error probably crept
back in with the big commit in early Jan 2011.  This commit fixes the issue when
compiling chem with KPP.  A tip of the hat goes to John Michalakes for all his 
help in fixing this problem.




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@4684 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d