Add the fix to Makefile and dyn_em/Makefile to eliminate an error message when compiling without chemistry.
Basically, we define a Makefile macro as the required files that we  
will need inside of start (with USEs) if there is a chem build (first  
mod in WRFV3/Makefile).  We pass that variable (second mod in WRFV3/ 
Makefile) into the dyn_em Makefile (this is the CF variable). It is  
either a blank or a list of those 4 files, based on the environment.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@4772 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d