Merging changes 5033-5140 of the HWRF_2011_oper branch to the HWRF pseudo-trunk of WRF.
Briefly,

1. The 2011 HWRF WRF emergency fixes are all in here, but

2. The SAS pgcon variable was used as tuning in the emergency 
fix.  The physically-based pre-emergency value of 0.55 is retained
as the default value in the code, and a new namelist variable
sas_pgcon is added to modify pgcon on a per-domain basis.  If
sas_pgcon is not specified (or if a non-NMM core is used), the
code will revert to the original value of 0.55.  The HWRF namelists
will have to set sas_pgcon=0.2 in &physics to use the tuning value
of pgcon.  All NMM cores have been modified to allow sas_pgcon to
be set, but I have not changed non-NMM cores.  

3. Three bug fixes from Ligia are added to correct the use of uninitialized
variables.


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/HWRF@5141 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d