Added a bug fix to my server polling version of the module_io_quilt. I had forgotten to add the code needed to get the server polling version of the I/O servers to exit, so they instead hung forever after the forecast finished.
You see, the master process of each I/O group does a blocking
MPI_SSend to the monitor process when the group is ready for a new
file.  When the monitor process has a file to write, it does an
MPI_Recv to MPI_ANY_SOURCE (any I/O server master processes) to find
one that is ready for a file.  However, at the end of the forecast,
when the monitor wants to quit, if it does not first post the MPI_Recv
to each I/O server, the I/O servers will hang forever once they finish
writing the last file.  This update adds the code to post those
receives so that the model will terminate correctly.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/stream1.5@5212 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d