Made changes necessary to get the HWRF stream1.5 branch to correctly reproduce the old 27:9 results.  These changes included:
1. The addition of a new -oldf configure option that will use the old optimization
exclusion settings, now in arch/noopt_exceptions_oldf

2. An #ifdef HWRF block that was missing from
dyn_nmm/module_NEST_UTIL.F is now added back in.

3. A missing IF statement in share/mediation_nest_move.F that has the
child move 6 gridpoints north if the vortex has moved 3 gridpoints
north.  This was removed in the new nest motion algorithms because it
leads to sporadic nest motion.  It has been added in, but is only activated
if the old HWRF motion algorithm is requested (vortex_tracker==1)

With this, the model will produce identical results to the current
HWRF pseudo-trunk if the correct namelist settings are given.  In
particular: vortex_tracker==1,1,0; terrain_smoothing==1; and
gfs_alpha=1,1.  Other namelist settings should be set the same as for
the old 27:9 HWRF.

Results will not be binary identical due to the new RANDOM variable in the 
history files.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/stream1.5@5213 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d