RUC LSM commit includes:
1. The input (and output) value of soil moisture is changed from (Q-Qmin) to Q, while internally to RUC LSM th
meaning of prognostic soil moisture variable remains the same -> (Q-Qmin). This change is made for consistency with
other LSM options in WRF and to avoid confusions in soil moisture initializations in research studies.

2. Added capability to use USGS or MODIS land use classifications. RUC LSM tables for MODIS classification were added 
to LANDUSE.TBL and VEGPARM.TBL. The choice of land-use classification is defined by namelist.wps for geogrid.

3. Added capability to specify soil and land-use parameters based on mosaic approach. Two new namelist parameters are 
introduced: mosaic_lu and mosaic_soil, the default values are 0, but to switch to mosaic specifications these namelist parameters
should be set to 1 (not necessary for both of them, could be default for soil and mosaic for land-use, or visa versa).

4. Tuning some parameters in snow model to get more realistic snow melting rates.

5. Several changes on the NMM side to make RUC LSM initialized correctly.




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@5238 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d