Corrected a bug in the HWRF coupler initialization, in the cases where coupling is unused.  The coupler timestep dtC still has to be set correctly.  Previously, when coupling is unused, dtC was set to the outer domain physics timestep.  It should be set to the domain 2 nest movement timestep.  This update corrects that.
Also added in this update is two new variables: dtc (coupler timestep),
and guessdtc, which is the domain 2 nest movement timestep.  The 
guessdtc is used as dtc in situations where uncoupled runs are done.

This code will still work if only one domain is used.  It will use the
outer domain nest movement timestep nphs*dt*movemin.  That is a 
somewhat nonsensical concept, as the outer domain (MOAD) does not move,
but it still has a movemin value so this does not matter.

This update will only impact the HWRF configuration.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/stream1.5@5253 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d