Merged revisions 5299 through 5318 from the stream1.5newphys branch to the stream1.5commit branch.  This adds a new version of the SAS shallow convection and the GFS PBL, taken from the 2011 GFS.  To use it, you must set sas_shal_conv=1, cu_physics=84 and bl_physics=83. These two upgrades must be used together or not at all since a PBL height array is passed between the two.  That means, if you set bl_physics=83, you must also set sas_shal_conv=1 and cu_physics=84, and vice-versa.  However, the cu_physics=84 option can still be used without bl_physics=83 if the sas_shal_conv is set to 0 (disabling shallow convection) and bl_physics is set to anything other than 83.
These changes are only applied to the NMM cores (all four NMM
configurations).  ARW still uses the old shallow convection.  It
should be easy to get the new shallow convection working in ARW if the
three new 2D arrays are added.  If someone does do that, I recommend
keeping support for the old shallow convection as well, and having a
new variable sas_shal_conv that can be set to 1 for the new shallow
convection, 0 for no shallow convection and something else for the old
shallow convection.  The 0 and 1 settings are needed to retain
consistency with NMM's sas_shal_conv.

ALSO in this update, the random number generation is disabled for
cu_physics=84 for both NMM and ARW (in module_cu_sas.F) since that
routine never actually used the random numbers it was given.  It
copied them from one array to another, and then promptly ignored them.

In NMM, turning on and off the use of the RANDOM array is now done
using in_use_for_config.  Currently, it is only needed by the OSAS
scheme, and so the Registry.NMM* now adds a
"state:randstate1,...,random" entry in the "package osasscheme" line.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/stream1.5commit@5319 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d