Updated the hwrf-baseline to the current WRF trunk, and included a bug fix from DTC needed to get the trunk to run without crashing.  In addition, another bug fix from Tim Brown is added to the NEST_TERRAIN to avoid invalid terrain heights.
If you run with the current HWRF physics, results will be binary
identical, but...

If you run with the new 2011 SAS shallow convection and 2011 PBL, then
results are not binary identical.  This is a somewhat worrying result
that I cannot explain.  

However, we need to start the baseline tests today, and the baseline
tests will not be using the 2011 GFS physics, so I'm merging these
changes.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/hwrf-baseline@5394 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d