UN-do fix.
1) Turns out that the "b" option for the dimension (meaning
that there is a boundary array) is NOT required when doing
lateral boundary forcing for the nest for, which is the "f"
flag.  This is not a bug fix, but removing an unnecessary
previous mod that makes the boundary file larger than
required. This is removed from these files:
M    Registry/Registry.NMM_HWRF
M    Registry/Registry.NMM_NEST
M    Registry/Registry.NMM_CHEM

2) The dfi_qnh mod is actually a bug in the Registry.  It should
be part of the dfi_scalar, not dfi_moist.  Unfortunately, somewhere
in the code there are bit-for-bit diffs with this change for 
NMM.  This change is being removed for all of the NMM Registry
files.  This is one of those "a bug is hiding a bug" scenarios.

M    Registry/Registry.NMM_HWRF
M    Registry/Registry.NMM_NEST
M    Registry/Registry.NMM_CHEM
M    Registry/Registry.NMM



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@5525 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d