Committed a species advected version of the HWRF Ferrier scheme.  This scheme is internally exactly the same as the non-advected version (and even shares the same code in phys/module_mp_HWRF.F) but the F_ICE, F_RIMEF and F_RAIN arrays no longer exist outside of the module functions in module_mp_HWRF.  Instead, QC, QI, QR and QNI are used to store those values and the F_* arrays are recalculated internally as needed.
However, this scheme, as with all other species advected schemes in the WRF-NMM, 
will be completely useless until nest-parent feedback of microphysics is implemented.
Currently, the WRF-NMM advects 0 condensate in all species at the boundary, 
leading to violent convection and turbulence at the nest boundary when a species
advected scheme is used.

In another branch, I am working on new interpolation functions to handle nest-parent
feedback of microphysics species.  Once that is complete, I will test it with
this branch's changes.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/hwrf-baseline-mpadvect@5529 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d