Disabled parallel make in WRF.
Dependencies in the WRF makefiles are wrong in many places, which is
causing occasional failed compiles on CCS due to files being compiled
in the wrong order.  We cannot risk NCO failing to build WRF due to
these incorrect dependencies, so I am disabling parallel make in the
hwrf-baseline branch.


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/hwrf-baseline@5581 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d