Reduced the number of variables included from "use module" statements in module_IGWAVE_ADJUST and removed the grid from one routine's parameter list.  This will not change the results of computation, but will allow the IBM compiler to compile the file again.  This change will likely speed compilation on other compilers as well.
This is needed to work around an IBM compiler bug.  An "upgrade" on the CCS operational 
machines in the past few days made the XLF compiler even less capable than it was before.
It could not optimize module_IGWAVE_ADJUST due to the large number of symbols in the 
scope.  I had fixed this before for bluefire by reducing the number of symbols in the
scope, but this CCS "upgrade" exacerbated XLF's problem further, necessitating a further
reduction in the number of symbols.

Summary of changes:

1. Don't include all of module_comm_dm.  Seriously, you should never do that; the
module is gigantic.  Only include what you need using the "only:" syntax.

2. Same for module_model_constants

3. Remove the unused "grid" parameter from VTOA's argument list.

4. Remove all "use module" statements from VTOA.


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/HWRF@5957 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d