Modified the Registry parser's interpolation/feedback/smoother/forcer calling code generation so that it can handle three new types of arguments to functions.  This new functionality will be needed by the new NMM interpolation routines.
The three new types of arguments:

integers -- f=(myfun:3016,avar,23,anothervar) -- you can pass
integers.  In this example, the function call would become:

    call myfun ( ... standard boundary interp args ...
                 , 3016 &
                 , grid%avar, ngrid%avar &
                 , 23 &
                 , grid%anothervar, ngrid%anothervar)

finite reals -- d=(myfun:13.24e-04,23.07,.05,somevar) -- POSIX style
real numbers (read man strtod for details) are re-expressed in
fortran-compliant scientific notation.  Infinite numbers such as NaN
or -Inf are not recognized.  Hexadecimal floats are recognized.
Anything that is recognized as a decimal integer is handled as an
integer instead, as described above.  This example would be:

    call myfun ( ... standard down interp args ...
                 , 13.24e-04 &
                 , 2.307e+01 &
                 , 5.0e-02 &
                 , grid%somevar, ngrid%somevar)

local variables -- u=(myfun:l%avar,gridvar) -- l%varname asks for the
local variable "varname" to be sent to the function.  In this example:

    call myfun ( ... standard up interp args ...
                 , avar &
                 , grid%gridvar, ngrid%gridvar)

To combine it all together:

    u=(myfun:l%local,3.7,100,gridvar)

becomes

    call myfun ( ... standard up interp args ...
                 , local &
                 , 3.7e+00 &
                 , 100 &
                 , grid%gridvar, ngrid%gridvar )



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/hwrf-newinterp@5965 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d