One of the boundary layer schemes (module_bl_boulac.F) has an inadvertent division by zero that was caught by the Intel compiler on Yellowstone.
Results from this calculation are not used and do not propagate as the code is written, but the developer agreed that it would be best if division by zero is avoided wherever possible.  
Because the variables "wts" and "wstar" are proportional, we can test the value of "wts".  

293c293,297
<         tstar_w=wts/wstar
---
>         if (wts .ne. 0.0) then
>            tstar_w=wts/wstar
>         else
>            tstar_w=0.0
>         endif

M    phys/module_bl_boulac.F



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@6225 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d