Provide consistency for both ARW em_real cases and idealized nested cases.
A previous fix to start_em.F corrected the base-state computation,
so that a single domain run and a nested domain (without feedback)
have the same reference state.  That fix basically swapped the
computation of t_init and alb.  

However, the order of that computation broke the idealized nested
cases.  Now we have an if test that put the computations in the
order required, depending on whether the case is idealized.

M    dyn_em/start_em.F



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@6341 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d