Comment from Greg Thompson:
That small IF-block check is something that is likely to have near zero effect on any model runs.
It basically will for homogeneous freezing of cloud water in the event the amount is between my R1 value (1.E-12 kg/kg) and the first entry found in my "r_c" array (1.E-6).  Since this only occurs below about -38 or -40C, it's not easy to get water there in the first place and it's for very small amount of cloud water.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@6436 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d