Put extra arrays for seaice albedo, seaice snowdepth, and seaice thickness into packages.  This got a little messy, trying to determine whether arrays are present or missing in the input file, and making sure the user doesn't choose namelist options that make use of fields that haven't been provided.
Also, change the default seaice albedo value in the namelist (option
SEAICE_ALBEDO_DEFAULT) from a fairly high 0.80 (good for heavy
Antarctic sea ice) to a more run-of-the-mill value of 0.65, more
likely to be "reasonable" for a random user's situation.

Add missing documentation (for SEAICE_ALBEDO_OPT=2) to the
README.namelist.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@6499 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d