>>>>>>>>>>WRFPLUS Branch Commit<<<<<<<<<<<<<<<
Contributed by Jonathan Guerrette (Jonathan Guerrette <Jonathan.Guerrette@Colorado.EDU>) 

A gradient validation code for the adjoint, tangent linear, and nonlinear finite difference.  This new code gives the capability to test the derivative between any two control variables (u,v,w,t,ph,mu,moist,tracer) and at any two locations in the domain.  Additional variables should not be too hard to add.  The user must select the combination of variables to be tested in the namelist, and create two files listing the coordinates of the points to be tested.  

This is an additional validation to the test that you already have in place, not a replacement.  The point is that using this method would enable the developer to determine the source of errors, as opposed to just knowing that they exist, and also to test sensitivities between multiple variables. 



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/WRFPLUSV3@6517 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d

Former-commit-id: 378d515edd9baec2d928d09d7b5acaf20ccb1284