This update fixes two bugs in the HWRF branch that prevent the model from running when floating point debugging is enabled:
1. There is a write statement in real_nmm.F that prints out a
partially uninitialized array of floats.  This was removed.

2. The start_domain_nmm runs the nest motion algorithm initialization
during real_nmm, which it should not do.  This was corrected by having
real_nmm set a special flag in start_domain_nmm to indicate that real
is being run.  Also, three variables had to be removed from the input
(i) stream.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/HWRF@6627 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d