Ugh, I hate finding these AFTER the release.
The released code goes out without either the chem or
the var subdirectories.  In this case, the chem directory
is explicitly assumed to exist as the top of the file has:

#define WRF_PORT
#define MODAL_AERO
#include "../chem/MODAL_AERO_CPP_DEFINES.h"

To fix this, we wrap an ifdef around the #include as
such:

#define WRF_PORT
#if defined ( WRF_CHEM )
#       include "../chem/MODAL_AERO_CPP_DEFINES.h"
#else
#       define MODAL_AERO
#       define MODAL_AERO_3MODE
#endif

This fixes errors when trying to compile about missing
include files.

M    phys/module_cam_mp_microp_aero.F



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@6665 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d